extract buildPackage into seperate files
