+++ packages/modular-scripts/src/build.ts
@@ -1,29 +1,35 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 
-import { outputDirectory, packagesRoot } from './config';
+import { outputDirectory } from './config';
 import getModularRoot from './utils/getModularRoot';
 import actionPreflightCheck from './utils/actionPreflightCheck';
 import isModularType from './utils/isModularType';
+import { getWorkspaceInfo } from './utils/getWorkspaceInfo';
 import execSync from './utils/execSync';
 
-async function build(
-  packagePath: string,
-  preserveModules?: boolean,
-): Promise<void> {
+async function build(target: string, preserveModules?: boolean): Promise<void> {
   const modularRoot = getModularRoot();
+  const workspace = await getWorkspaceInfo();
+  const record = workspace[target];
+  if (!record) {
+    throw new Error(`${target} does not exist in modular workspace`);
+  }
 
-  if (isModularType(path.join(modularRoot, packagesRoot, packagePath), 'app')) {
+  const targetPath = path.join(modularRoot, record.location);
+
+  if (isModularType(targetPath, 'app')) {
     const buildScript = require.resolve(
       'modular-scripts/react-scripts/scripts/build.js',
     );
 
     // create-react-app doesn't support plain module outputs yet,
     // so --preserve-modules has no effect here
-    await fs.remove(path.join(outputDirectory, packagePath));
+    await fs.remove(path.join(outputDirectory, target));
+
     // TODO: this shouldn't be sync
     execSync('node', [buildScript], {
-      cwd: path.join(modularRoot, packagesRoot, packagePath),
+      cwd: targetPath,
       log: false,
       // @ts-ignore
       env: {
@@ -32,14 +38,14 @@ async function build(
     });
 
     await fs.move(
-      path.join(packagesRoot, packagePath, 'build'),
-      path.join(outputDirectory, packagePath),
+      path.join(targetPath, 'build'),
+      path.join(outputDirectory, target),
     );
   } else {
     const { buildPackage } = await import('./buildPackage');
     // ^ we do a dynamic import here to defer the module's initial side effects
     // till when it's actually needed (i.e. now)
-    await buildPackage(packagePath, preserveModules);
+    await buildPackage(target, preserveModules);
   }
 }
