+++ packages/modular-scripts/src/config/index.ts
@@ -1,2 +1,1 @@
-export const packagesRoot = 'packages';
 export const outputDirectory = 'dist';
